Fix execution of integration tests with host keys
Fixed integration tests failing due to not providing a host key
callback to the SSH target creation function. For now, this has been
worked around by simply ignoring host keys in the tests.

The "correct" way of fixing this would be to have the different servers
output predictable host keys, but that would be much more complicated.