Move majority of implementation to internal
To reduce the complexity of the primary ex package, most of the code
that existed in that directory has been split out and moved to an
internal directory.

This places most of the SSH code behind abstractions, allowing it to
more easily be changed later and to reduce the external interface of the
ex package for ease of use and lessened noise.